+++ examples/source/src/main.rs
@@ -3,9 +3,7 @@ use once_cell::sync::Lazy;
 
 static ENV: Lazy<Environment<'static>> = Lazy::new(|| {
     let mut env = Environment::new();
-    let mut source = Source::new();
-    source.load_from_path("templates", &["txt"]).unwrap();
-    env.set_source(source);
+    env.set_source(Source::from_path("templates"));
     env
 });